refactor(ai): consolidate Responses helpers and tests
